move ga file from optimizer to util
